requirements.txt
apache-beam==2.28.0
avro-python3==1.9.2.1
bcrypt==3.2.0
bigquery==0.0.12
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crcmod==1.7
cryptography==3.4.6
dacktool==0.0.7
dbstream==0.0.30
dill==0.3.1.1
docopt==0.6.2
fastapi==0.63.0
fastavro==1.3.2
fasteners==0.16
future==0.18.2
google-api-core==1.26.0
google-api-python-client==1.7.11
google-apitools==0.5.31
google-auth==1.27.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
google-cloud-bigquery
google-cloud-bigquery-storage==2.3.0
google-cloud-bigtable==1.7.0
google-cloud-build==2.0.0
google-cloud-core==1.6.0
google-cloud-datastore==1.15.3
google-cloud-dlp==1.0.0
google-cloud-language==1.3.0
google-cloud-pubsub==1.7.0
google-cloud-spanner==1.19.1
google-cloud-videointelligence==1.16.1
google-cloud-vision==1.0.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
googleauthentication==0.0.15
grpc-google-iam-v1==0.12.3
grpcio==1.36.0
grpcio-gcp==0.2.2
gunicorn==20.0.4
h11==0.12.0
h2==4.0.0
hdfs==2.6.0
hpack==4.0.0
httplib2==0.17.4
Hypercorn==0.11.2
hyperframe==6.0.0
idna==2.10
libcst==0.3.17
mock==2.0.0
mypy-extensions==0.4.3
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.3
paramiko==2.7.2
pbr==5.5.1
priority==1.3.0
proto-plus==1.14.2
protobuf==3.15.3
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8
pydot==1.4.2
pymongo==3.11.3
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sshtunnel==0.1.5
starlette==0.13.6
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.26.3
wsproto==1.0.0
